category plot. default colors for categorygraphics
