+++ plugin/jvm/src/main/java/com/twosigma/beaker/chart/serializer/XYChartSerializer.java
@@ -31,21 +31,6 @@ import org.codehaus.jackson.map.SerializerProvider;
  */
 public class XYChartSerializer extends AbstractChartSerializer<XYChart> {
 
-  private static class ColorPalette {
-    static final Color[] colors = new Color[]{
-      new Color(192, 80, 77),
-      new Color(79, 129, 189),
-      new Color(155, 187, 89),
-      new Color(247, 150, 70),
-      new Color(128, 100, 162),
-      new Color(75, 172, 198)
-    };
-
-     static Color getColor(int i) {
-      return colors[i % colors.length];
-    }
-  }
-
   @Override
   public void serialize(XYChart xychart, JsonGenerator jgen, SerializerProvider sp)
     throws IOException, JsonProcessingException {
